# be/src/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "API SinapSense",
    "version": "1.0.0",
    "description": "API de integração de dados",
    "contact": {
      "name": "Lucas José de Souza",
      "email": "souza.lucasj@ufpr.br"
    }
  },
  "servers": [
    {
      "url": "/",
      "description": "API Base URL"
    }
  ],

  "components": {
    "schemas": {
      "File": {
        "type": "object",
        "properties": {
          "filename": {
            "type": "string",
            "description": "Nome do arquivo"
          },
          "description": {
            "type": "string",
            "description": "Descrição do arquivo"
          },
          "path": {
            "type": "string",
            "description": "Caminho do arquivo"
          },
          "active": {
            "type": "boolean",
            "default": true,
            "description": "Status de ativação do arquivo"
          },
          "jsonData": {
            "type": "object",
            "description": "Dados em formato JSON do arquivo"
          }
        },
        "required": ["filename", "path", "jsonData"]
      }
    }
  },

  "paths": {
    "/sign-in": {
      "post": {
        "summary": "Realiza login do usuário",
        "description": "Autentica o usuário com email e senha, retornando um token de acesso JWT.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email",
                    "example": "usuario@email.com"
                  },
                  "password": {
                    "type": "string",
                    "format": "password",
                    "example": "minhasenha123"
                  }
                },
                "required": ["email", "password"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login realizado com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "accessToken": {
                      "type": "string",
                      "description": "Token JWT para autenticação"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Erro de validação nos dados",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "path": {
                        "type": "array",
                        "items": { "type": "string" }
                      },
                      "message": {
                        "type": "string"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Credenciais inválidas",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Invalid credentials."
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Erro interno no servidor"
          }
        },
        "tags": ["auth"]
      }
    },
    "/sign-up": {
      "post": {
        "summary": "Cria uma nova conta de usuário",
        "description": "Registra um novo usuário com nome, email e senha.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "example": "Pedro da Silva"
                  },
                  "email": {
                    "type": "string",
                    "format": "email",
                    "example": "pedrinho@email.com"
                  },
                  "password": {
                    "type": "string",
                    "format": "password",
                    "example": "minhasenha123"
                  }
                },
                "required": ["name", "email", "password"]
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Conta criada com sucesso (sem conteúdo na resposta)"
          },
          "400": {
            "description": "Erro de validação nos dados",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "path": {
                        "type": "array",
                        "items": { "type": "string" }
                      },
                      "message": {
                        "type": "string"
                      }
                    }
                  }
                }
              }
            }
          },
          "409": {
            "description": "Email já está em uso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "This email is already in use."
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Erro interno no servidor"
          }
        },
        "tags": ["auth"]
      }
    },
    "/heatmap": {
      "post": {
        "summary": "Faz upload de um arquivo CSV e um arquivo de mídia para gerar o heatmap JSON",
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "csvFile": {
                    "type": "string",
                    "format": "binary",
                    "description": "Arquivo CSV que será processado para gerar o heatmap"
                  },
                  "mediaFile": {
                    "type": "string",
                    "format": "binary",
                    "description": "Arquivo de mídia relacionado ao experimento (JPEG, PNG, MP4, etc.)"
                  },
                  "filename": {
                    "type": "string",
                    "description": "Nome do experimento"
                  },
                  "description": {
                    "type": "string",
                    "description": "Descrição do experimento"
                  },
                  "mediaType": {
                    "type": "integer",
                    "enum": [0, 1],
                    "description": "Tipo da mídia: 0 = imagem, 1 = vídeo"
                  }
                },
                "required": ["file", "filename"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Arquivos processados e dados salvos com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Arquivo CSV e mídia processados e salvos com sucesso!"
                    },
                    "data": {
                      "$ref": "#/components/schemas/File"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Erro de validação na requisição",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Nome do arquivo é obrigatório"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Erro interno no servidor",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Erro ao processar o arquivo CSV"
                    },
                    "details": {
                      "type": "string",
                      "example": "Entre em contato com o administrador"
                    }
                  }
                }
              }
            }
          }
        },
        "tags": ["upload"]
      }
    },
    
    "/heatmap/{id}": {
      "put": {
        "summary": "Atualiza metadados e mídia de um heatmap existente",
        "description": "Permite atualizar o nome, descrição e imagem de um arquivo de heatmap existente.",
        "tags": ["update"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "ID do heatmap a ser atualizado"
          }
        ],
        "requestBody": {
          "required": false,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "filename": {
                    "type": "string",
                    "description": "Novo nome do experimento (opcional)"
                  },
                  "description": {
                    "type": "string",
                    "description": "Nova descrição do experimento (opcional)"
                  },
                  "mediaFile": {
                    "type": "string",
                    "format": "binary",
                    "description": "Nova mídia para substituir a existente (opcional)"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Arquivo atualizado com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Arquivo atualizado com sucesso!"
                    },
                    "data": {
                      "$ref": "#/components/schemas/File"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Arquivo não encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Arquivo não encontrado"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Erro interno no servidor",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Erro ao atualizar o arquivo"
                    },
                    "details": {
                      "type": "string",
                      "example": "Detalhes do erro interno"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },

    "/eyetracking": {
      "get": {
        "summary": "Retorna todos os registros de eyetracking ativos",
        "responses": {
          "200": {
            "description": "Lista de arquivos de eyetracking ativos",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/File"
                  }
                }
              }
            }
          }
        },
        "tags": ["list"]
      }
    },
    "/eyetracking/{_id}": {
      "get": {
        "summary": "Retorna um registro de eyetracking pelo ID",
        "parameters": [
          {
            "name": "_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Dados do arquivo retornados com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/File"
                }
              }
            }
          },
          "404": {
            "description": "Arquivo não encontrado"
          },
          "400": {
            "description": "ID inválido"
          }
        },
        "tags": ["list"]
      },
      "put": {
        "summary": "Atualiza a visibilidade de um registro de eyetracking",
        "parameters": [
          {
            "name": "_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "active": {
                    "type": "boolean"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Status do arquivo atualizado com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/File"
                }
              }
            }
          },
          "404": {
            "description": "Arquivo não encontrado"
          },
          "400": {
            "description": "ID inválido"
          }
        },
        "tags": ["update"]
      }
    }
  }
}